+++ indico/modules/events/registration/controllers/management/regforms.py
@@ -399,13 +399,14 @@ class RHRegistrationFormModify(RHManageRegFormBase):
         min_data_retention = data_retention_settings.get('minimum_data_retention')
         max_data_retention = data_retention_settings.get('maximum_data_retention') or timedelta(days=3650)
         regform_retention_weeks = self.regform.retention_period.days // 7 if self.regform.retention_period else None
+        has_predefined_affiliations = Affiliation.query.filter_by(is_deleted=False).has_rows()
         return WPManageRegistration.render_template('management/regform_modify.html', self.event,
                                                     form_data=get_flat_section_setup_data(self.regform),
                                                     regform=self.regform,
                                                     data_retention_range={'min': min_data_retention.days // 7,
                                                                           'max': max_data_retention.days // 7,
                                                                           'regform': regform_retention_weeks},
-                                                    has_predefined_affiliations=Affiliation.query.has_rows())
+                                                    has_predefined_affiliations=has_predefined_affiliations)
 
 
 class RHRegistrationFormStats(RHManageRegFormBase):