+++ CHANGES.rst
@@ -16,6 +16,8 @@ Improvements
   buttons for guests and users who do not have access to them (:pr:`7332`, thanks
   :user:`jbtwist, unconventionaldotdev`)
 - Add option to filter room booking calendar to show only unused rooms (:pr:`7350`)
+- Add "Affiliation" field type to regforms (:pr:`7352`, thanks
+  :user:`duartegalvao, unconventionaldotdev`)
 
 Bugfixes
 ^^^^^^^^